requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dealabs-api @ git+https://github.com/IDerr/dealabs-api.git@e2b5a1058f4f93e82327c5ff53964dcdd279eeae
fastapi==0.116.0
Flask==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
pillow==11.3.0
playwright==1.54.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-multipart==0.0.20
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
yarl==1.20.1
flask
gunicorn
# autres dépendances ici
